# Conflicts:
#	com.vrcfury.vrcfury/Editor/VF/Feature/ToggleBuilder.cs

+++ com.vrcfury.vrcfury/Editor/VF/PlayModeTrigger.cs
@@ -1,43 +1,44 @@
 using System;
-using System.IO;
 using System.Linq;
-using System.Reflection;
 using UnityEditor;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 using VF.Builder;
 using VF.Builder.Exceptions;
 using VF.Component;
 using VF.Inspector;
 using VF.Menu;
 using VF.Model;
-using VF.PlayMode;
 using VF.Service;
-using VRC.Dynamics;
+using VF.Utils;
 using VRC.SDK3.Avatars.Components;
 using VRC.SDKBase.Editor.BuildPipeline;
 using Object = UnityEngine.Object;
 
 namespace VF {
-    public class PlayModeTrigger {
+    internal class PlayModeTrigger {
         private static string tmpDir;
+        private const string TriggerObjectName = "__vrcf_play_mode_trigger";
+        private static bool scannedThisFrame = false;
 
         [InitializeOnLoadMethod]
-        static void Init() {
-            SceneManager.sceneLoaded += OnSceneLoaded;
+        private static void Init() {
             EditorApplication.playModeStateChanged += OnPlayModeStateChanged;
+            VRCFuryComponent._OnValidate = () => {
+                if (Application.isPlaying && !addedTriggerObjectThisPlayMode) {
+                    addedTriggerObjectThisPlayMode = true;
+                    var obj = new GameObject(TriggerObjectName);
+                    RescanOnStartComponent.AddToObject(obj, true);
+                }
+            };
+            Scheduler.Schedule(() => scannedThisFrame = false, 0);
         }
 
+        private static bool addedTriggerObjectThisPlayMode = false;
         private static void OnPlayModeStateChanged(PlayModeStateChange state) {
             if (state == PlayModeStateChange.ExitingEditMode) {
                 tmpDir = null;
+                addedTriggerObjectThisPlayMode = false;
             }
-        }
-
-        private static void OnSceneLoaded(Scene scene, LoadSceneMode mode) {
-            if (VrcIsUploadingDetector.IsProbablyUploading()) return;
-            if (!EditorApplication.isPlaying) return;
-            Rescan(scene);
         }
 
         // This should absolutely always be false in play mode, but we check just in case
@@ -56,19 +57,22 @@ namespace VF {
             return false;
         }
 
-        private static void Rescan(Scene scene) {
+        private static void Rescan() {
             if (!Application.isPlaying) return;
             if (!PlayModeMenuItem.Get()) return;
+            if (scannedThisFrame) return;
+            scannedThisFrame = true;
 
             if (tmpDir == null) {
                 var tmpDirParent = TmpFilePackage.GetPath() + "/PlayMode";
                 VRCFuryAssetDatabase.DeleteFolder(tmpDirParent);
                 tmpDir = $"{tmpDirParent}/{DateTime.Now.ToString("yyyyMMdd-HHmmss")}";
-                Directory.CreateDirectory(tmpDir);
+                VRCFuryAssetDatabase.CreateFolder(tmpDir);
             }
 
-            foreach (var root in VFGameObject.GetRoots(scene)) {
+            foreach (var root in VFGameObject.GetRoots()) {
                 foreach (var avatar in root.GetComponentsInSelfAndChildren<VRCAvatarDescriptor>()) {
+                    if (avatar == null) continue; // it was deleted
                     RescanOnStartComponent.AddToObject(avatar.owner());
                     var obj = avatar.owner();
                     if (!obj.activeInHierarchy) continue;
@@ -114,7 +118,10 @@ namespace VF {
                     VRCFExceptionUtils.ErrorDialogBoundary(() => {
                         try {
                             var hapticContactsService = new HapticContactsService();
-                            VRCFuryHapticPlugEditor.Bake(plug, hapticContactsService, tmpDir);
+                            var bakeResult = VRCFuryHapticPlugEditor.Bake(plug, hapticContactsService, tmpDir);
+                            foreach (var renderer in bakeResult.renderers) {
+                                SaveAssetsBuilder.SaveUnsavedComponentAssets(renderer.renderer, tmpDir);
+                            }
                         } catch (Exception e) {
                             throw new ExceptionWithCause($"Failed to bake detached SPS Plug: {plug.owner().GetPath()}", e);
                         }
@@ -136,13 +143,18 @@ namespace VF {
         [DefaultExecutionOrder(-10000)]
         public class RescanOnStartComponent : MonoBehaviour {
             private void Start() {
-                Rescan(this.owner().scene);
+                Rescan();
+                var obj = gameObject;
+                DestroyImmediate(this);
+                if (obj.name == TriggerObjectName) {
+                    DestroyImmediate(obj);
+                }
             }
 
-            public static void AddToObject(VFGameObject obj) {
-                if (obj.GetComponent<RescanOnStartComponent>()) {
-                    return;
-                }
+            public static void AddToObject(VFGameObject obj, bool evenIfAlreadyEnabled = false) {
+                if (!Application.isPlaying) return;
+                if (obj.GetComponent<RescanOnStartComponent>() != null) return;
+                if (!evenIfAlreadyEnabled && obj.activeInHierarchy) return;
                 obj.AddComponent<RescanOnStartComponent>();
             }
         }
